refactor(cli): migrate radar tooling onto discovery
Move the NEXRAD radar refresher/writer/validator/live_validator (the NOAA HOMR
fixed-width parser and the unsigned chunk-feed lister) to earthlens.radar.cli,
dropping core's last direct earthlens.radar imports. Relocate the CLI tests, keeping
the shared _get_text test in core. Shrink both guard allow-lists (radar).
