+++ libs/core/tests/cli/test_refresh_helpers.py
@@ -21,46 +21,8 @@ def _info(provider):
     return next(b for b in list_backends() if b.provider == provider)
 
 
-class TestRadarParse:
-    """Tests for the HOMR fixed-width radar parser + writer."""
-
-    _COLS = [("ICAO", 5), ("NAME", 11), ("ST", 3), ("LAT", 10), ("LON", 11)]
-
-    def _homr(self):
-        """Build a tiny fixed-width HOMR table aligned to the dash rule."""
-
-        def line(vals):
-            return " ".join(v.ljust(w) for v, (_, w) in zip(vals, self._COLS))
-
-        rule = " ".join("-" * w for _, w in self._COLS)
-        header = line([name for name, _ in self._COLS])
-        rows = [
-            line(["KABR", "ABERDEEN", "SD", "45.4558", "-98.4131"]),
-            line(["KXXX", "BADLAT", "ZZ", "999.0", "0.0"]),
-            line(["AB", "SHORT", "XX", "1.0", "2.0"]),
-            line(["K1AB", "NUM", "XX", "1.0", "2.0"]),
-            line(["KZZZ", "NOLAT", "XX", "abc", "2.0"]),
-        ]
-        return "\n".join([header, rule, *rows])
-
-    def test_keeps_only_valid_stations(self):
-        """Only 4-letter alphabetic ICAO rows with in-range coords survive."""
-        rows = refresh_mod._radar_station_rows(self._homr())
-        assert list(rows) == ["KABR"], rows
-        assert rows["KABR"]["name"] == "Aberdeen", "name title-cased"
-        assert rows["KABR"]["state"] == "SD" and rows["KABR"]["latitude"] == 45.4558
-
-    def test_short_table_is_empty(self):
-        """A table with fewer than three lines yields no rows."""
-        assert refresh_mod._radar_station_rows("only\ntwo") == {}
-
-    def test_no_icao_column_is_empty(self):
-        """A table whose header lacks ICAO yields no rows."""
-        assert refresh_mod._radar_station_rows("FOO BAR\n--- ---\nx   y") == {}
-
-    def test_station_ids_sorted(self):
-        """_radar_station_ids returns the sorted ICAO ids only."""
-        assert refresh_mod._radar_station_ids(self._homr()) == ["KABR"]
+class TestGetText:
+    """Tests for the shared `_get_text` HTTP helper."""
 
     def test_get_text_returns_body(self, monkeypatch):
         """_get_text returns the response body text."""
@@ -72,16 +34,6 @@ class TestRadarParse:
             ),
         )
         assert refresh_mod._get_text("https://x") == "BODY"
-
-    def test_write_radar_rewrites_stations(self, tmp_path, monkeypatch):
-        """_write_radar re-parses HOMR into the curated stations: block."""
-        monkeypatch.setattr(refresh_mod, "_get_text", lambda url: self._homr())
-        target = tmp_path / "radar_data_catalog.yaml"
-        target.write_text("stations:\n  KOLD:\n    name: Old\n", encoding="utf-8")
-        monkeypatch.setattr(refresh_mod, "_index_path", lambda info: target)
-        path = refresh_mod._write_radar(_info("radar"), {"radar": ["KABR"]})
-        data = yaml.safe_load(open(path))
-        assert "KABR" in data["stations"], "HOMR rows written"
 
 
 class TestChcWalk:
